Fix broken access

+++ app/views.py
@@ -1,7 +1,7 @@
 import sqlite3
 from django.shortcuts import render, redirect
 from django.contrib.auth import login, authenticate, logout
-#from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseForbidden
 from django.views.decorators.csrf import csrf_exempt
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
@@ -64,18 +64,21 @@ def create_view(request):
     return render(request, "create.html", {"form": form})
 
 @csrf_exempt
+@login_required()
 def delete_view(request, post_id):
-    #if request.user != post.user:
-    #return HttpResponseForbidden()
     post = Post.objects.get(id=post_id)
-    post.delete()
+    if request.user != post.author:
+        return HttpResponseForbidden()
+    else:
+        post.delete()
     return redirect("home")
 
 @csrf_exempt
+@login_required()
 def edit_view(request, post_id): 
-    #if request.user != post.user:
-    #   return HttpResponseForbidden()
     post = Post.objects.get(id=post_id)
+    if request.user != post.user:
+       return HttpResponseForbidden()
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
